move `space serve` to `space dev serve`
