+++ cmd/dev/root.go
@@ -90,6 +90,7 @@ The cli will start one process for each of your micros, then expose a single enp
 	cmd.AddCommand(newCmdDevUp())
 	cmd.AddCommand(newCmdDevProxy())
 	cmd.AddCommand(newCmdDevTrigger())
+	cmd.AddCommand(newCmdServe())
 
 	cmd.Flags().StringP("dir", "d", ".", "directory of the project")
 	cmd.Flags().StringP("id", "i", "", "project id")
@@ -329,7 +330,7 @@ func MicroCommand(micro *types.Micro, directory, projectKey string, port int) (*
 		if root == "" {
 			root = micro.Src
 		}
-		devCommand = fmt.Sprintf("%s serve %s --port %d", shellescape.Quote(os.Args[0]), shellescape.Quote(root), port)
+		devCommand = fmt.Sprintf("%s dev serve %s --port %d", shellescape.Quote(os.Args[0]), shellescape.Quote(root), port)
 	} else if EngineToDevCommand[micro.Engine] != "" {
 		devCommand = EngineToDevCommand[micro.Engine]
 	} else {
